Improve performance on windows
Change opencv backend API to DSHOW, and changes related to this, which allows for higher resolution and native camera options window. OpenCV camera support on windows is apparently atrocious, but this had the best performance on Win 10, python 3.10, and OpenCV 4.x.

Also, initial code for automatic camera checking to try to find the proper index. WIP.